# docs/rest_v1_openapi.yaml
openapi: 3.0.2
servers:
  - url: http://localhost:3370
    description: Local Linstor controller
info:
  description: |
    Linstor REST API V1

    The V1 rest api of Linstor should stay compatible and only additions are made to the API,
    If there are breaking changes or redesigned a new major REST API version will be issued.

    Server runs per default on port `3370` on `::` ipv6 and ipv4.

    To change the bind address or port you can use the following linstor client commands:
    ```
    linstor controller set-property REST/bindAddress 127.0.0.1
    linstor controller set-property REST/port 8080
    ```

    After setting this properties restart the controller and the new values should be used.

    Changelog:

    * unreleased
       - Added makeResourceAvailable API
       - Added diskless storage pool filter to AutoSelectFilter
       - Added external_locking to StoragePool
    * 1.6.0
       - Added PropsInfoAll, PropsInfo and PropsInfoDetails as well as these entrypoints:
          - /v1/controller/properties/info
          - /v1/controller/properties/info/all
          - /v1/nodes/properties/info
          - /v1/storage-pool-definitions/properties/info
          - /v1/nodes/{node}/storage-pools/properties/info
          - /v1/resource-definitions/properties/info
          - /v1/resource-definitions/{resource}/resources/properties/info
          - /v1/resource-definitions/{resource}/volume-definitions/properties/info
...
       - Added /v1/nodes/{node}/restore
...
paths:
...
  /v1/physical-storage:
    get:
      tags:
        - developers
      operationId: viewPhysicaStorage
      summary: show physical storage on nodes, to be used with LINSTOR
      description: Gives a grouped list of physical storage that can be turned into
        a LINSTOR storage-pool
      parameters:
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: physical storage list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PhysicalStorage'
        '500':
          $ref: '#/components/responses/OperationFailed'
  /v1/physical-storage/{node}:
    post:
      tags:
        - developers
      operationId: createDevicePool
      summary: creates an LVM, LVM-thin or ZFS pool, optional VDO under it
      description: |
        Creates a LVM/LVM-thin, ZFS pool on the given device and if supported VDO(optional) under it.
        logcal_size_kib parameter is only needed if LVM-thin or vdo is used.
        Also note VDO can only used with LVM-fat.
      parameters:
        - $ref: '#/components/parameters/Node'
      responses:
        '201':
          description: device pool created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiCallRcList'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhysicalStorageCreate'
            example:
              provider_kind: LVM_THIN
              device_path: /dev/sde
              pool_name: ssd_fast
              logcal_size_kib: 5368709120
              vdo_enable: false
...
components:
  parameters:
...
    Node:
      name: node
      in: path
      required: true
      description: node to use
      schema:
        type: string
...
    Offset:
      name: offset
      in: query
      description: number of records to skip for pagination
      schema:
        type: integer
        format: int32
        minimum: 0
    Limit:
      name: limit
      in: query
      description: maximum number of records to return
      schema:
        type: integer
        format: int32
        minimum: 0
        maximum: 50
...
  responses:
...
    OperationFailed:
      description: operation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiCallRc'
  schemas:
    ApiCallRc:
      description: Common api reply structure
      type: object
      required:
        - ret_code
        - message
      properties:
        ret_code:
          type: integer
          format: int64
          description: A masked error number
        message:
          type: string
          example: Operation result.
        cause:
          type: string
          description: Cause of the error
        details:
          type: string
          description: Details to the error message
        correction:
          type: string
          description: Possible correction options
        error_report_ids:
          type: array
          description: List of error report ids related to this api call return code.
          items:
            type: string
        obj_refs:
          type: object
          description: Map of objection that have been involved by the operation.
          additionalProperties:
            type: string
    ApiCallRcList:
      type: array
      items:
        $ref: '#/components/schemas/ApiCallRc'
...
    Properties:
      type: object
      description: A string to string property map.
      additionalProperties:
        type: string
    ProviderKind:
      type: string
      enum:
        - DISKLESS
        - LVM
        - LVM_THIN
        - ZFS
        - ZFS_THIN
        - OPENFLEX_TARGET
        - FILE
        - FILE_THIN
        - SPDK
...
    PhysicalStorage:
      type: object
      properties:
        size:
          type: integer
          format: int64
        rotational:
          type: boolean
        nodes:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/PhysicalStorageDevice'
    PhysicalStorageDevice:
      type: object
      properties:
        device:
          type: string
        model:
          type: string
        serial:
          type: string
        wwn:
          type: string
    PhysicalStorageStoragePoolCreate:
      type: object
      description: This structure is used for create physical-storage
      properties:
        name:
          type: string
          description: Name of the linstor storage pool
        props:
          $ref: '#/components/schemas/Properties'
    PhysicalStorageCreate:
      type: object
      description: If `with_storage_pool` is set a linstor storage pool will also
        be created using this device pool
      required:
        - provider_kind
        - device_paths
      properties:
        provider_kind:
          $ref: '#/components/schemas/ProviderKind'
        device_paths:
          type: array
          items:
            type: string
        raid_level:
          type: string
          description: |
            RAID level to use for pool.
          enum:
            - JBOD
          default: JBOD
        pool_name:
          type: string
        vdo_enable:
          default: false
          type: boolean
        vdo_slab_size_kib:
          type: integer
          format: int64
          default: 0
        vdo_logical_size_kib:
          type: integer
          format: int64
          default: 0
        with_storage_pool:
          $ref: '#/components/schemas/PhysicalStorageStoragePoolCreate'
...